+++ src/cli/input_notes.rs
@@ -295,7 +295,7 @@ mod tests {
     use miden_client::{
         config::{ClientConfig, Endpoint},
         errors::NoteIdPrefixFetchError,
-        mock::{mock_full_chain_mmr_and_notes, mock_notes, MockClient, MockDataStore, MockRpcApi},
+        mock::{mock_full_chain_mmr_and_notes, mock_notes, MockClient, MockRpcApi},
         store::{sqlite_store::SqliteStore, InputNoteRecord},
     };
     use miden_lib::transaction::TransactionKernel;
@@ -317,11 +317,12 @@ mod tests {
         );
 
         let store = SqliteStore::new((&client_config).into()).unwrap();
+        let executor_store = SqliteStore::new((&client_config).into()).unwrap();
 
         let mut client = MockClient::new(
             MockRpcApi::new(&Endpoint::default().to_string()),
             store,
-            MockDataStore::default(),
+            executor_store,
         )
         .unwrap();
 
@@ -366,11 +367,12 @@ mod tests {
             Endpoint::default().into(),
         );
         let store = SqliteStore::new((&client_config).into()).unwrap();
+        let executor_store = SqliteStore::new((&client_config).into()).unwrap();
 
         let mut client = MockClient::new(
             MockRpcApi::new(&Endpoint::default().to_string()),
             store,
-            MockDataStore::default(),
+            executor_store,
         )
         .unwrap();
 
@@ -397,11 +399,12 @@ mod tests {
         );
 
         let store = SqliteStore::new((&client_config).into()).unwrap();
+        let executor_store = SqliteStore::new((&client_config).into()).unwrap();
 
         let mut client = MockClient::new(
             MockRpcApi::new(&Endpoint::default().to_string()),
             store,
-            MockDataStore::default(),
+            executor_store,
         )
         .unwrap();